requirements.txt
torch
torchvision
pytest
Pillow
pyyaml
matplotlib
pytest-html
numpy
setuptools 